% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ds2ISOCurrency}
\alias{ds2ISOCurrency}
\title{change currency datastream code into iso version}
\usage{
ds2ISOCurrency(dsCurrency)
}
\arguments{
\item{dsCurrency}{The datastream currency code}
}
\value{
the iso code for the currency
}
\description{
change currency datastream code into iso version
}

