% Generated by roxygen2 (4.0.0): do not edit by hand
\name{EncryptPassword}
\alias{EncryptPassword}
\title{Encrypt the Datastream password}
\usage{
EncryptPassword(strPassword = "")
}
\arguments{
\item{strPassword}{the password to be encrypted}
}
\value{
an encrypted password
}
\description{
Encrypt the Datastream password
}
\details{
This is a port of the VBA code
}

