% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getDataStream}
\alias{getDataStream}
\title{Initialise connection with Datastream DWE server}
\usage{

  getDataStream(dweURLwsdl = "http://dataworks.thomson.com/Dataworks/Enterprise/1.0/webServiceClient.asmx?WSDL",
  User = as.character("USERNAME"), Pass = as.character("PASSWORD"))
}
\arguments{
\item{dweURLwsdl}{The URL of the server}

\item{User}{The username for Datastream.  If not provided it will use the username from the windows registry}

\item{Pass}{The password for Datastream.  Also sourced from Registry}
}
\value{
a DWE connection object
}
\description{

}
\details{
\code{getDataStream} initialises the connection with the Datastream DWE server
}

