% Generated by roxygen2 (4.0.0): do not edit by hand
\name{iso2DatastreamCurrency}
\alias{iso2DatastreamCurrency}
\title{change currency ISO code into Datastream version}
\usage{
iso2DatastreamCurrency(isoCurrency, multiplier = 1)

iso2DatastreamCurrency(isoCurrency, multiplier = 1)
}
\arguments{
\item{isoCurrency}{The iso currency code to be converted to the Datastream version}

\item{isoCurrency}{The iso currency code to be converted to the Datastream version}

\item{multiplier}{An addition multiplier to key using}
}
\value{
the Datastream code for the currency

the Datastream code for the currency
}
\description{


The use of the multiplier parameter alongs minor divisions to be obtained eg pence and cents
}

