% Generated by roxygen2 (4.0.0): do not edit by hand
\name{staticRequest}
\alias{staticRequest}
\title{make a static request

\code{staticRequest} Function that returns a the value of Expression for the array of instruments in DSCode from Datastream
parameters are}
\usage{
staticRequest(dwei = getDataStream(), DSCode, Expression = "",
  endDate = Sys.Date(), frequency = "D", verbose = FALSE)
}
\arguments{
\item{dwei}{- A Datastream Client Interface object created with getDataStream}

\item{DSCode}{- an array of instruments eg c("RIO","MKS")}

\item{Expression}{- the data to return eg MNEM or NAME}

\item{startDate}{- the start date of the timeseries}

\item{endDate}{- the end date of the timeseries}

\item{frequency}{- the frequency of the request}

\item{verbose}{- whether to give messages during the request}
}
\value{
returns an array of the requested information
}
\description{
make a static request

\code{staticRequest} Function that returns a the value of Expression for the array of instruments in DSCode from Datastream
parameters are
}

